small changes to miner.cpp for building

+++ src/libgpusolver/libgpusolver.h
@@ -66,7 +66,7 @@ class GPUSolver {
 
 public:
 	GPUSolver();
-	GPUSolver(int64_t selGPU);
+	GPUSolver(unsigned selGPU);
 	~GPUSolver();
         bool run(unsigned int n, unsigned int k, uint8_t *header, size_t header_len, uint64_t nonce,
 		            const std::function<bool(std::vector<unsigned char>)> validBlock,
